requirements.txt
python-telegram-bot==20.8
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

